Install packages via vcpkg
In this commit, all of the windows-base, build vcpkg cache and
windows-ci tasks are synced to install packages via vcpkg in end Windows
ci image. The end image includes the dependencies that are installed by
both vcpkg and installers.

To use dependencies that are installed by vcpkg:

1- Remove extra_[lib|include]_dirs for the openssl, otherwise there will
be a conflict.

2- PG_DEPS_['PATH', 'PKG_CONFIG', 'PKG_CONFIG_PATH',
  'EXTRA_INCLUDE_DIRS', 'EXTRA_LIB_DIRS'] are exported as global env
  variablles. Use them accordingly, for example:
    * PATH: ${PG_DEPS_PATH}${PATH}
    * PKG_CONFIG: ${PG_DEPS_PKG_CONFIG}
    * PKG_CONFIG_PATH: ${PG_DEPS_PKG_CONFIG_PATH}
    * -Dextra_include_dirs=%PG_DEPS_EXTRA_INCLUDE_DIRS%
    * -Dextra_lib_dirs=%PG_DEPS_EXTRA_LIB_DIRS%

Summary of the code changes:

- windows_install_packages_via_vcpkg.ps1 script is modified to be called
  from both windows.pkr.hcl and build vcpkg cache task.

- windows-ci task depends on build-vcpkg-cache task now.